requirements.txt
pytest>=7.4.2
pre-commit>=2.15.0
PyYAML>=6.0
